Cleaning up some older tests that are not relevant
