+++ tests/sam_api_test.py
@@ -12,32 +12,25 @@ class SAMAPITestCase(unittest.TestCase):
     def tearDown(self):
         pass
 
-    # def test_requests_retry_session(self):
-    #     ops = get_opps.get_opps_for_day(limit=100, filter_naics=True)
-    #     self.assertGreater(len(ops), 10)
-    #
-    #     for opp in ops:
-    #         if not any(opp['naicsCode'].startswith(n) for n in sam_utils.naics_code_prefixes):
-    #             self.assertTrue(False, f"Failure - opp was returned that had naics code {opp['naicsCode']}. Only naics codes in {','.join(sam_utils.naics_code_prefixes)} should be returned")
-    #
-    #
-    # def test_schematize_opps(self):
-    #     opps = get_opps.get_opps_for_day(limit=100, filter_naics=True)
-    #     op = opps[0]
-    #     schematized_opp = get_opps.schematize_opp(op)
-    #
-    #     keys = ("noticeId", "title", "solicitationNumber", "agency", "office", "type", "naicsCode", "psc", "classificationCode", "pointOfContact")
-    #     for k in keys:
-    #         self.assertIn(k, schematized_opp)
-    #
-    # def test_transform(self):
-    #     opps = get_opps.get_opps_for_day(limit=10, filter_naics=True)
-    #     get_opps.transform_opps(opps, "/tmp/")
+    def test_requests_retry_session(self):
+        ops = get_opps.get_opps_for_day(limit=100, filter_naics=True)
+        self.assertGreater(len(ops), 10)
 
-    def test_update_old_solicitations(self):
-        conn_string = get_db_url()
-        dal = DataAccessLayer(conn_string)
-        dal.connect()
+        for opp in ops:
+            if not any(opp['naicsCode'].startswith(n) for n in sam_utils.naics_code_prefixes):
+                self.assertTrue(False, f"Failure - opp was returned that had naics code {opp['naicsCode']}. Only naics codes in {','.join(sam_utils.naics_code_prefixes)} should be returned")
 
-        with session_scope(dal) as session:
-            sam_utils.update_old_solicitations(session, age_cutoff=365, max_tests=5)
+
+    def test_schematize_opps(self):
+        opps = get_opps.get_opps_for_day(limit=100, filter_naics=True)
+        op = opps[0]
+        schematized_opp = get_opps.schematize_opp(op)
+
+        keys = ("noticeId", "title", "solicitationNumber", "agency", "office", "type", "naicsCode", "psc", "classificationCode", "pointOfContact")
+        for k in keys:
+            self.assertIn(k, schematized_opp)
+
+    def test_transform(self):
+        opps = get_opps.get_opps_for_day(limit=10, filter_naics=True)
+        get_opps.transform_opps(opps, "/tmp/")
+
